mqtt2serial/requirements.txt
paho-mqtt
pyserial
Pillow
